requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
rapidfuzz==3.6.1
python-dotenv==1.0.0